change: refactoring sample
