+++ gemini-android/src/main/java/com/haroncode/gemini/android/binder/RestoreBinder.kt
@@ -75,5 +75,4 @@ internal class RestoreBinder<View : SavedStateRegistryOwner>(
             ConnectionRulesFactoryManager.clear(name)
         }
     }
-
 }
