+++ sample/src/main/java/com/haroncode/gemini/sample/presentation/routing/MainConnectionFactory.kt
@@ -1,7 +1,6 @@
 package com.haroncode.gemini.sample.presentation.routing
 
-import com.haroncode.gemini.android.connector.DelegateConnectionRulesFactory
-import com.haroncode.gemini.android.connector.connectionRulesFactory
+import com.haroncode.gemini.android.binder.DelegateConnectionRulesFactory
 import com.haroncode.gemini.connector.ConnectionRulesFactory
 import com.haroncode.gemini.sample.di.scope.PerFragment
 import com.haroncode.gemini.sample.ui.MainFragment
