+++ sample/src/main/java/com/haroncode/gemini/sample/ui/AuthFragment.kt
@@ -6,7 +6,7 @@ import android.widget.Toast
 import androidx.core.view.isVisible
 import androidx.core.widget.doOnTextChanged
 import com.haroncode.gemini.StoreEventListener
-import com.haroncode.gemini.android.connector.StoreViewBinder
+import com.haroncode.gemini.android.binder.StoreViewBinding
 import com.haroncode.gemini.sample.R
 import com.haroncode.gemini.sample.base.PublisherFragment
 import com.haroncode.gemini.sample.databinding.FragmentAuthBinding
@@ -16,13 +16,14 @@ import com.haroncode.gemini.sample.presentation.onlyaction.AuthStore.Action
 import com.haroncode.gemini.sample.presentation.onlyaction.AuthStore.Action.LoginClick
 import com.haroncode.gemini.sample.presentation.onlyaction.AuthStore.State
 import javax.inject.Inject
+import javax.inject.Provider
 
 class AuthFragment :
     PublisherFragment<Action, State>(R.layout.fragment_auth),
     StoreEventListener<AuthStore.Event> {
 
     @Inject
-    lateinit var factory: AuthConnectionFactory
+    lateinit var factory: Provider<AuthConnectionFactory>
 
     private var _binding: FragmentAuthBinding? = null
     private val binding get() = _binding!!
@@ -30,7 +31,7 @@ class AuthFragment :
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        StoreViewBinder.of(shouldSave = false) { factory }
+        StoreViewBinding.withRestore(factoryProvider = factory::get)
             .bind(this)
     }
