+++ sample/src/main/java/com/haroncode/gemini/sample/ui/MainFragment.kt
@@ -2,7 +2,7 @@ package com.haroncode.gemini.sample.ui
 
 import android.os.Bundle
 import android.view.View
-import com.haroncode.gemini.android.connector.StoreViewBinder
+import com.haroncode.gemini.android.binder.StoreViewBinding
 import com.haroncode.gemini.sample.R
 import com.haroncode.gemini.sample.base.PublisherFragment
 import com.haroncode.gemini.sample.databinding.FragmentMainBinding
@@ -20,7 +20,7 @@ class MainFragment : PublisherFragment<MainStore.Action, Unit>(R.layout.fragment
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        StoreViewBinder.of { factory }
+        StoreViewBinding.withRestore { factory }
             .bind(this)
     }
